LugaresSeguros/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
Pillow==9.2.0
pipenv==2022.1.8
platformdirs==2.5.1
psycopg2==2.8.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.1
virtualenv-clone==0.5.7
